requirements.txt
torchcodec
Flask
flask-cors
pytest
torch
torchaudio
librosa
numpy
faiss-cpu
torchopenl3
soundfile
tqdm
six
julius
nnAudio
pandas
